Change test file structure
